+++ packages/widget-reorder/src/control.tsx
@@ -1,7 +1,35 @@
 import React from 'react'
-import styled from '@emotion/styled'
-import { DragDropContext, Droppable, Draggable, DropResult } from 'react-beautiful-dnd'
+import { createPortal } from 'react-dom'
+import { fromJS, List, Map as ImmutableMap } from 'immutable'
+import { DropResult } from 'react-beautiful-dnd'
+import { WidgetProps } from '@ncwidgets/common-typings'
+import { normalize, diff, reorder } from './utils'
+import {
+  ControlList, 
+  ControlDraggableItem, 
+} from './components/drag-drop'
+import {
+  Modal, 
+  Modified, 
+  EmptyMessage,
+} from './components/modal'
 import { parseTemplate } from './utils'
+
+
+interface PreviewPortalProps {
+  portalRef: React.RefObject<HTMLDivElement>;
+  children: React.ReactNode;
+}
+
+export const PreviewPortal = ({portalRef, children}: PreviewPortalProps) => 
+  portalRef && portalRef.current && createPortal(children, portalRef.current)
+
+export interface RenderControlProps {
+  value: any;
+  field: ImmutableMap<any, any>;
+}
+
+export type RenderControl = (props: RenderControlProps) => React.ReactNode;
 
 export const renderDefaultControl = ({ value, field }) => {
   const fieldId = field.get('id_field')
@@ -9,160 +37,130 @@ export const renderDefaultControl = ({ value, field }) => {
   return parseTemplate({ template, data: value })
 }
 
-interface StyledDraggableProps {
-  isDragging: boolean;
+export interface ControlProps extends WidgetProps {
+  value: List<string>;
 }
 
-const StyledDraggable = styled.div<StyledDraggableProps>`
-  padding: 1rem;
-  opacity: ${({ isDragging }) => isDragging ? 0.6 : 1};
-  box-shadow: ${({ isDragging }) => isDragging ? '0 4px 16px 0 rgba(0,0,0,0.2)' : '0 2px 6px 0 rgba(0,0,0,0.2)'};
-  background: #fff;
-  border-radius: 3px;
-  margin-bottom: 0.5rem;
-`
-
-export const ControlDraggableItem = ({ identifier, index, children }) =>
-  <Draggable
-    key={identifier}
-    draggableId={identifier}
-    index={index}>
-    {(provided, snapshot) => (
-      <StyledDraggable
-        isDragging={snapshot.isDragging}
-        ref={provided.innerRef}
-        {...provided.draggableProps}
-        {...provided.dragHandleProps}
-        style={provided.draggableProps.style}
-      >
-        {children}
-      </StyledDraggable>
-    )}
-  </Draggable>
-
-
-interface StyledDroppableProps {
-  isDraggingOver: boolean;
-  maxHeight: string;
-}
-
-const StyledDroppable = styled.div<StyledDroppableProps>`
-  position: relative;
-  padding: 1rem;
-  background-color: ${({ isDraggingOver }) => isDraggingOver ? 'lightblue' : '#dfdfe3'};
-  border-radius: 3px;
-  overflow-y: scroll;
-  max-height: ${({ maxHeight }) => maxHeight};
-  border: 2px solid #dfdfe3;
-`
-
-export const ControlList = ({ onDragEnd, children, maxHeight }) => 
-  <DragDropContext onDragEnd={(result: DropResult) => onDragEnd(result)}>
-    <Droppable droppableId="droppable">
-      {(provided, snapshot) => (
-        <StyledDroppable
-          ref={provided.innerRef}
-          maxHeight={maxHeight}
-          isDraggingOver={snapshot.isDraggingOver}
-          {...provided.droppableProps}
-        >
-          { children }
-          { provided.placeholder }
-        </StyledDroppable>
-      )}
-    </Droppable>
-  </DragDropContext>
-
-
-const StyledHeader = styled.h1`
-  font-size: 1.25rem;
-  line-height: 1.3;
-  margin-bottom: 1rem;
-`
-
-const StyledModal = styled.div`
-  position: absolute;
-  width: 50%;
-  padding: 1rem 1rem 1.25rem;
-  text-align: center;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%,-50%);
-  background-color: #fff;
-  border-radius: 0.25rem;
-  box-shadow: 0 6px 12px 0 rgba(0,0,0,0.2);
-
-  button {
-    border: 1px solid #dfdfe3;
-    color: rgb(23, 162, 184);
-    padding: 0.75em 1.25em;
-    background: #fff;
-    border-radius: 6px;
-  }
-
-  button:hover {
-    border: 1px solid rgb(23, 162, 184);
-    background: rgb(23, 162, 184);
-    color: #fff;
-  }
-`
-
-const StyledOverlay = styled.div`
-  position: absolute;
-  width: 100%;
-  height: 100%;
-  background-color: rgba(223, 223, 227, 0.8);
-  z-index: 100;
-`
-
-export type Modified = 'none' | 'unset' | 'modified'
-
-interface ModalContentArgs {
-  collection: string;
+export interface ControlState {
+  data: Record<string, any>;
+  fetched: boolean;
+  newOrder: string[];
   modified: Modified;
 }
 
-type ModalContent = Record<'title' | 'action', string>
+export const createControl = ({ 
+  renderControl,
+  renderPreview,
+  previewRef
+}) => {
+  return class Control extends React.Component<ControlProps, ControlState> {
+    public state: ControlState = {
+      data: {},
+      fetched: false,
+      newOrder: [],
+      modified: 'none'
+    }
 
-const getModalContent = ({ modified, collection}: ModalContentArgs): ModalContent => {
-  let title = ''
-  let action = ''
+    public async componentDidMount() {
+      const { forID, field, value, query } = this.props
+      const collection: string = field.get('collection')
+      const fieldId: string = field.get('id_field')
 
-  if (modified === 'unset') {
-    title = `Order for collection '${collection}' hasn\'t been set yet`
-    action = 'Start Ordering'
+      // no value or empty value, assuming value is always a List<string>
+      const noValue = (typeof value === 'undefined' || value.size === 0) 
+
+      const result = await query(forID, collection, [fieldId], '')
+      const sourceData = result.payload.response.hits.map(payload => payload.data)
+      // @ts-ignore
+      const normalizedData = normalize(sourceData, fieldId)
+      this.setState({ data: normalizedData, fetched: true })
+
+      const idData: string[] = sourceData.map(item => item[fieldId])
+
+      if (noValue) {
+        this.setState({ newOrder: idData, modified: 'unset' })
+        return
+      }
+
+      const currentOrder = value.toJS()
+      const { newOrder, modified } = diff({
+        currentOrder,
+        data: idData,
+      })
+    
+      if (modified) {
+        this.setState({ newOrder, modified: 'modified' })
+      }
+    }
+
+    public handleDisplayChange = () => {
+      const { newOrder } = this.state
+      const { onChange } = this.props
+      this.setState({ modified: 'none' }, () => {
+        onChange(fromJS(newOrder))
+      })
+    }
+
+    public handleDragEnd = (result: DropResult) => {
+      if (
+        !result.destination || 
+      result.source.index === result.destination.index
+      ) return
+
+      const { value, onChange } = this.props
+      const data = value.toJS()
+
+      const sortedData = reorder({
+        data,
+        startIndex: result.source.index,
+        endIndex: result.destination.index
+      })
+
+      onChange(fromJS(sortedData))
+    }
+
+    public render() {
+      const { field, value, classNameWrapper } = this.props
+      const { modified, data, fetched } = this.state
+
+      const collection: string = field.get('collection')
+      const maxHeight: string = field.get('max_height', 'none')
+
+      const noValue = (typeof value === 'undefined' || value.size === 0) 
+
+      if (!fetched) return <div>loading...</div>
+      if (fetched && Object.keys(data).length === 0) {
+        return (
+          <EmptyMessage className={classNameWrapper} collection={collection} />
+        )
+      }
+      return (
+        <div style={{ position: 'relative', minHeight: '12rem' }}>
+          {modified !== 'none' && <Modal handleDisplayChange={this.handleDisplayChange} {...{ collection, modified }} />}
+          {!noValue && <ControlList onDragEnd={this.handleDragEnd} maxHeight={maxHeight}>
+            {
+              value.map((id, i) => {
+                const item = data[id]
+                const value = (typeof item === 'undefined')
+                  ? { error: 'Entry does not exist' }
+                // @ts-ignore
+                  : item
+                return (
+                  <ControlDraggableItem key={id} identifier={id} index={i}>
+                    {renderControl({ value, field })}
+                  </ControlDraggableItem>)
+              })
+            }
+          </ControlList>}
+        
+          { /* Renders preview in splitpane via this component... */ }
+          <PreviewPortal portalRef={previewRef}>
+            {!noValue && renderPreview({
+              value: value.map(identifier => data[identifier])
+            }) }
+          </PreviewPortal>
+        </div>
+      )
+    }
   }
-  if (modified === 'modified') {
-    title = `Collection '${collection}' have been changed`
-    action = 'Apply changes'
-  }
-  return { title, action }
 }
-
-interface ModalProps extends ModalContentArgs {
-  handleDisplayChange: () => void;
-}
-
-export const Modal: React.FC<ModalProps> = ({ collection, modified, handleDisplayChange }) => {
-  const { title, action } = getModalContent({ modified, collection })
-  return (
-    <StyledOverlay>
-      <StyledModal>
-        <StyledHeader>{title}</StyledHeader>
-        <button onClick={() => handleDisplayChange()}>{action}</button>
-      </StyledModal>
-    </StyledOverlay>
-  )
-}
-
-const StyledEmptyMessage = styled.div`
-  padding: 1rem;
-  text-align: center;
-`
-
-export const EmptyMessage = ({ className, collection }) => (
-  <StyledEmptyMessage className={className}>
-    <StyledHeader>Collection &apos;{collection}&apos; is empty</StyledHeader>
-    <p><a href={`#/collections/${collection}/new`} target="blank">Create new entries</a></p>
-  </StyledEmptyMessage>
-)
